Add new icon in text field
