+++ Attendance_App/Attendance_App/ViewControllers/HomeView/HomeViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class InfoViewController: UIViewController {
+class HomeViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
