+++ Attendance_App/Attendance_App/ViewControllers/LoginView/Notification.swift
@@ -1,12 +1,9 @@
 //
-//  SetUpComponents.swift
+//  Notification.swift
 //  Attendance_App
 //
-//  Created by J.K on 4/26/21.
+//  Created by J.K on 4/27/21.
 //  Copyright © 2021 J.K. All rights reserved.
 //
 
 import Foundation
-import UIKit
-
-
