+++ Attendance_App/Attendance_App/ViewControllers/LoginView/SetUpComponents.swift
@@ -1,0 +1,10 @@
+//
+//  SetUpComponents.swift
+//  Attendance_App
+//
+//  Created by J.K on 4/26/21.
+//  Copyright © 2021 J.K. All rights reserved.
+//
+
+import Foundation
+import UIKit
